Move filterOverview page into its own folder
